Move async logic to data in api
